Stats mapping rework, Statistics controller & StatsHelper logic

+++ RTA-Project-BL/Models/ArticleBL.cs
@@ -6,7 +6,7 @@ namespace RTA_Project_BL.Models
     public class ArticleBL
     {
         public int Id { get; set; }
-        public byte LangId { get; set; }
+        public int LangId { get; set; }
         public string Title { get; set; }
         public string Content { get; set; }
         public string AuthorId { get; set; }
