+++ RTA-Project-BL/Models/HeroBL.cs
@@ -3,7 +3,7 @@
     public class HeroBL
     {
         public int Id { get; set; }
-        public byte FactionId { get; set; }
+        public int FactionId { get; set; }
         public string NameRus { get; set; }
         public string NameEng { get; set; }
     }
