+++ RTA-Project-BL/Models/TournamentGroupBL.cs
@@ -5,7 +5,7 @@ namespace RTA_Project_BL.Models
     public class TournamentGroupBL
     {
         public int Id { get; set; }
-        public byte GroupFormatId { get; set; }
+        public int GroupFormatId { get; set; }
         public string NameRus { get; set; }
         public string NameEng { get; set; }
