+++ RTA-Project-DAL/Models/Hero.cs
@@ -6,7 +6,7 @@ namespace RTA_Project_DAL.Models
     public class Hero
     {
         public int Id { get; set; }
-        public byte FactionId { get; set; }
+        public int FactionId { get; set; }
         public string NameRus { get; set; }
         public string NameEng { get; set; }
         [DefaultValue(false)]
